removed depreciated inputs code
