--- runSettings.txt
@@ -1,22 +1,0 @@
-# String Valued Aegis Run Parameters
-DAGMC_input [s] inres1.h5m
-VTK_input [s] inres1.stl
-runcase [s] eqdsk
-launchPos [s] fixed
-trace [s] yes
-eqdsk_file [s] eqdsk/EQ3.eqdsk 
-
-# Integer Valued Aegis Run Parameters
-cenopt [i] 2
-number_of_rays_launched_per_tri [i] 5
-nTrack [i] 100000
-
-# Double Valued Aegis Run Parameters
-dsTrack [d] 0.05
-Psol [d] 7.5e+06
-lambda_q [d] 0.012
-psiref [d] -2.1628794
-rOutrBdry [d] 8.07385
-rmove [d] -0.006
-# Vector<int> Valued Aegis Run Parameters
-surfs [v] 3
