--- settings.txt
@@ -1,13 +1,0 @@
-DAGMC_input EQ3_cubes.h5m
-VTK_input EQ3_cubes.stl
-ray_qry exps/exps00010000.qry
-runcase eqdsk
-trace yes
-eqdsk_file eqdsk/equil_MAST.eqdsk 
-Psol 3.2
-lambda_q 6.187
-cenopt 2
-surfs 0
-number_of_rays_launched_per_tri 5
-nTrack 100000
-dsTrack 0.05
