+++ test/data/aegis_settings.json
@@ -1,0 +1,27 @@
+{
+  "description": "inres1 test case to compare against SMARDDA",
+  "aegis_params":{
+      "DAGMC": "inres1_shad.h5m",
+      "step_size": 0.005,
+      "max_steps": 100000,
+      "launch_pos": "fixed",
+      "target_surfs": [1,2,3,4,5,6],
+      "force_no_deposition": false
+  },
+  "equil_params":{
+      "eqdsk": "EQ3.eqdsk",
+      "cenopt": 2,
+      "P_sol": 7.5e+06,
+      "lambda_q": 0.012,
+      "psiref": -2.1628794,
+      "r_outrbdry": 8.07385,
+      "rmove": -0.006,
+      "draw_equil_rz": true,
+      "draw_equil_xyz": true,
+      "print_debug_info": false
+  },
+  "vtk_params":{
+      "VTK": "target_facets.stl",
+      "draw_particle_tracks": true
+  }
+ }
